wip: separate vouchers and contributions in new product api
